🚚 move: AuthorProfile 컴포넌트 공통 컴포넌트 폴더로 이동
